Move "Util.java" into its own package
