+++ src/ecgberht/BaseLocationComparator.java
@@ -1,6 +1,7 @@
 package ecgberht;
 
 import bwem.Base;
+import ecgberht.Util.Util;
 import org.openbw.bwapi4j.Position;
 
 import java.util.Comparator;
