+++ src/ecgberht/BehaviourTrees/Defense/CheckPerimeter.java
@@ -5,7 +5,7 @@ import ecgberht.EnemyBuilding;
 import ecgberht.GameState;
 import ecgberht.Squad;
 import ecgberht.Squad.Status;
-import ecgberht.Util;
+import ecgberht.Util.Util;
 import org.iaie.btree.state.State;
 import org.iaie.btree.task.leaf.Conditional;
 import org.iaie.btree.util.GameHandler;
