+++ src/ecgberht/BehaviourTrees/Training/ChooseMedic.java
@@ -1,7 +1,7 @@
 package ecgberht.BehaviourTrees.Training;
 
 import ecgberht.GameState;
-import ecgberht.Util;
+import ecgberht.Util.Util;
 import org.iaie.btree.state.State;
 import org.iaie.btree.task.leaf.Action;
 import org.iaie.btree.util.GameHandler;
