+++ src/ecgberht/BuildingMap.java
@@ -4,6 +4,7 @@ import bwem.BWEM;
 import bwem.Base;
 import bwem.ChokePoint;
 import bwem.area.Area;
+import ecgberht.Util.Util;
 import org.openbw.bwapi4j.BW;
 import org.openbw.bwapi4j.Player;
 import org.openbw.bwapi4j.Position;
