+++ src/ecgberht/EnemyBuilding.java
@@ -1,5 +1,6 @@
 package ecgberht;
 
+import ecgberht.Util.Util;
 import org.openbw.bwapi4j.TilePosition;
 import org.openbw.bwapi4j.type.UnitType;
 import org.openbw.bwapi4j.unit.Building;
