+++ src/ecgberht/GameState.java
@@ -11,6 +11,7 @@ import com.google.gson.Gson;
 import ecgberht.Agents.*;
 import ecgberht.Simulation.SimManager;
 import ecgberht.Strategies.*;
+import ecgberht.Util.Util;
 import jfap.JFAP;
 import org.iaie.btree.util.GameHandler;
 import org.openbw.bwapi4j.*;
