+++ src/ecgberht/IntelligenceAgency.java
@@ -3,6 +3,7 @@ package ecgberht;
 import ecgberht.Strategies.BioBuild;
 import ecgberht.Strategies.BioBuildFE;
 import ecgberht.Strategies.BioMechBuildFE;
+import ecgberht.Util.Util;
 import org.openbw.bwapi4j.Bullet;
 import org.openbw.bwapi4j.type.Race;
 import org.openbw.bwapi4j.type.UnitType;
